#1126: unify the parallel iterator state-machine stub emitters
The six reference-type iterator creation stubs — free-function / instance /
static, once for the sync generator and again for the async generator — emitted
byte-identical IL, differing only in the static `smBuilder` type (four even
carried the literal comment "Mirrors EmitAsyncStubMethod"). Fold them into two
shared emitters:

* `EmitIteratorFreeFunctionStub` — newobj → snapshot the thread-local dynamic
  `this` into `<>4__this` (#775) → copy object-typed params → seed the function
  display class → return.
* `EmitIteratorMethodStub(…, bool isInstanceMethod, string? funcDCKey)` — the
  instance/static method form, merged the way `EmitAsyncStubMethod` already merges
  its two variants: `this` from arg 0 for instance methods (params then at arg 1),
  value-type params boxed into the object fields, function DC seeded only when a
  key is supplied (static iterator methods pass null, preserving their existing
  no-write-capture behaviour).

The static builder type is abstracted behind a new `IIteratorStateMachineBuilder`
interface (Constructor / ThisField / FunctionDCField / GetVariableField), which
both `GeneratorStateMachineBuilder` and `AsyncGeneratorStateMachineBuilder` already
satisfy. The async *function* stub stays separate on purpose: its state machine is
a value-type struct initialised via ldloca/initobj with an AsyncTaskMethodBuilder —
a different shape, not a copy.

Pure code move — byte-identical IL. Build green; generator/async-generator/iterator
targeted suite 1469/1469.
