+++ Compilation/ILCompiler.Generators.cs
@@ -309,7 +309,7 @@ public partial class ILCompiler
             var methodBuilder = _functions.Builders[funcName];
 
             // Emit the stub method body (creates and returns the state machine)
-            EmitGeneratorStubMethod(methodBuilder, smBuilder, funcStmt, funcName);
+            EmitIteratorFreeFunctionStub(methodBuilder, smBuilder, funcStmt, funcName);
 
             // Emit the MoveNext method body
             EmitGeneratorMoveNextBody(smBuilder, funcStmt, funcName);
@@ -321,68 +321,10 @@ public partial class ILCompiler
         _currentNamespacePath = savedNamespacePath;
     }
 
-    /// <summary>
-    /// Emits the stub method that creates and initializes the generator state machine.
-    /// </summary>
-    private void EmitGeneratorStubMethod(
-        MethodBuilder methodBuilder,
-        GeneratorStateMachineBuilder smBuilder,
-        Stmt.Function funcStmt,
-        string qualifiedName)
-    {
-        var il = methodBuilder.GetILGenerator();
-
-        // Create new instance of the state machine using the constructor builder
-        il.Emit(OpCodes.Newobj, smBuilder.Constructor);
-
-        // #775: capture the active dynamic `this` into the state machine's <>4__this field. The stub
-        // runs eagerly (when the generator is created), while MoveNext runs lazily — by then the
-        // thread-local receiver is gone, so it must be snapshotted here. The receiver is the thread-local
-        // `$TSFunction._currentFunctionThis` (set by InvokeWithThis for an `o.gen()` / `.call(recv)`
-        // value-call), coerced null → globalThis sentinel to match LocalVariableResolver.LoadThis.
-        if (smBuilder.ThisField != null && _runtime?.CurrentFunctionThisField != null)
-        {
-            il.Emit(OpCodes.Dup);       // Keep state machine reference on stack
-            il.Emit(OpCodes.Ldsfld, _runtime.CurrentFunctionThisField);
-            if (_runtime.GlobalThisSingletonField != null)
-            {
-                var thisNotNull = il.DefineLabel();
-                il.Emit(OpCodes.Dup);
-                il.Emit(OpCodes.Brtrue, thisNotNull);
-                il.Emit(OpCodes.Pop);
-                il.Emit(OpCodes.Ldsfld, _runtime.GlobalThisSingletonField);
-                il.MarkLabel(thisNotNull);
-            }
-            il.Emit(OpCodes.Stfld, smBuilder.ThisField);
-        }
-
-        // Copy parameters to state machine fields
-        for (int i = 0; i < funcStmt.Parameters.Count; i++)
-        {
-            var paramName = funcStmt.Parameters[i].Name.Lexeme;
-            var field = smBuilder.GetVariableField(paramName);
-            if (field != null)
-            {
-                il.Emit(OpCodes.Dup);  // Keep state machine reference on stack
-                il.Emit(OpCodes.Ldarg, i);
-                il.Emit(OpCodes.Stfld, field);
-            }
-        }
-
-        // Instantiate the function display class (#674) and seed any captured-and-mutated
-        // parameters into it so an arrow that writes them shares the generator's storage. The
-        // stub params are object-typed (BuildStateMachineStubParamTypes), so no boxing is needed.
-        EmitGeneratorFunctionDCInit(il, smBuilder.FunctionDCField, funcStmt, qualifiedName, paramOffset: 0);
-
-        // Captured outer-scope variables are NOT copied into the state machine here. Doing so
-        // snapshotted their value at creation time, so a later mutation of the outer variable
-        // was invisible to the running body (#541). MoveNext instead reads/writes them live
-        // from their enclosing storage (top-level statics / entry-point display class), which
-        // requires the corresponding fields to be set on the MoveNext CompilationContext below.
-
-        // Return the state machine (which implements IEnumerable<object>)
-        il.Emit(OpCodes.Ret);
-    }
+    // EmitGeneratorStubMethod (free function), EmitGeneratorInstanceStubMethod, and
+    // EmitGeneratorStaticStubMethod were folded into the shared EmitIteratorFreeFunctionStub /
+    // EmitIteratorMethodStub (ILCompiler.IteratorStubs.cs, #1126) — byte-identical with the async
+    // generator's three copies.
 
     /// <summary>
     /// With the state machine instance on the stack, news up the function display class (#674),
@@ -541,10 +483,7 @@ public partial class ILCompiler
         // (#692) has no function-DC write-capture support (it is not registered in
         // RegisterGeneratorMethodFunctionDisplayClasses, so methodDCKey is null) and a write-capture
         // inside one still fail-fasts safely via the CapturedWriteAnalysis guard.
-        if (isInstanceMethod)
-            EmitGeneratorInstanceStubMethod(methodBuilder, smBuilder, method, methodDCKey);
-        else
-            EmitGeneratorStaticStubMethod(methodBuilder, smBuilder, method.Parameters);
+        EmitIteratorMethodStub(methodBuilder, smBuilder, method, isInstanceMethod, isInstanceMethod ? methodDCKey : null);
 
         // Create context for MoveNext emission
         var il = smBuilder.MoveNextMethod.GetILGenerator();
@@ -623,111 +562,6 @@ public partial class ILCompiler
         smBuilder.CreateType();
     }
 
-    /// <summary>
-    /// Emits the stub method that creates and initializes the generator state machine for an instance method.
-    /// The stub copies 'this' and parameters to the state machine, then returns it.
-    /// </summary>
-    private void EmitGeneratorInstanceStubMethod(
-        MethodBuilder methodBuilder,
-        GeneratorStateMachineBuilder smBuilder,
-        Stmt.Function method,
-        string? funcDCKey)
-    {
-        var parameters = method.Parameters;
-        var il = methodBuilder.GetILGenerator();
-
-        // Create new instance of the state machine
-        il.Emit(OpCodes.Newobj, smBuilder.Constructor);
-
-        // Copy 'this' to state machine's ThisField if it exists
-        if (smBuilder.ThisField != null)
-        {
-            il.Emit(OpCodes.Dup);  // Keep state machine reference on stack
-            il.Emit(OpCodes.Ldarg_0);  // Load 'this' (instance methods have 'this' at arg 0)
-            il.Emit(OpCodes.Stfld, smBuilder.ThisField);
-        }
-
-        // Box value types since state machine fields are object-typed. Decide from the method's ACTUAL
-        // IL signature (methodBuilder.GetParameters()), not the AST-resolved types: a private method's
-        // parameters are all `object` slots, so boxing the AST-resolved value type (e.g. Double) would
-        // mismatch the `object` argument actually loaded (StackUnexpected). Mirrors EmitAsyncStubMethod.
-        var paramTypes = methodBuilder.GetParameters();
-
-        // Copy parameters to state machine fields (instance methods start params at index 1)
-        for (int i = 0; i < parameters.Count; i++)
-        {
-            var paramName = parameters[i].Name.Lexeme;
-            var field = smBuilder.GetVariableField(paramName);
-            if (field != null)
-            {
-                il.Emit(OpCodes.Dup);  // Keep state machine reference on stack
-                il.Emit(OpCodes.Ldarg, i + 1);  // +1 because 'this' is at index 0
-
-                if (i < paramTypes.Length && paramTypes[i].ParameterType.IsValueType)
-                {
-                    il.Emit(OpCodes.Box, paramTypes[i].ParameterType);
-                }
-
-                il.Emit(OpCodes.Stfld, field);
-            }
-        }
-
-        // #724: instantiate the function display class and seed any captured-and-mutated parameter
-        // into it so an arrow that writes that parameter shares the generator's storage. Instance
-        // methods carry 'this' at arg 0, so user params start at arg 1 (paramOffset: 1). The stub
-        // params are typed, so EmitGeneratorFunctionDCInit boxes value types before the store. No-op
-        // when the method has no function DC.
-        if (funcDCKey != null)
-            EmitGeneratorFunctionDCInit(il, smBuilder.FunctionDCField, method, funcDCKey, paramOffset: 1, paramTypes);
-
-        // Captured outer-scope variables are NOT copied into the state machine (#541): see the
-        // free-function stub above. MoveNext reads/writes them live from their backing storage,
-        // wired through the CompilationContext in EmitGeneratorMethodBody.
-
-        // Return the state machine (which implements IEnumerable<object>)
-        il.Emit(OpCodes.Ret);
-    }
-
-    /// <summary>
-    /// Emits the stub that creates the generator state machine for a STATIC method (#692): like the
-    /// instance stub but with no <c>this</c> and parameters starting at arg 0 (no receiver slot).
-    /// Value-type parameters are boxed into the object-typed state-machine fields.
-    /// </summary>
-    private void EmitGeneratorStaticStubMethod(
-        MethodBuilder methodBuilder,
-        GeneratorStateMachineBuilder smBuilder,
-        List<Stmt.Parameter> parameters)
-    {
-        var il = methodBuilder.GetILGenerator();
-
-        // Create new instance of the state machine
-        il.Emit(OpCodes.Newobj, smBuilder.Constructor);
-
-        // Box value types since state machine fields are object-typed. Decide from the method's ACTUAL
-        // IL signature (methodBuilder.GetParameters()), not the AST-resolved types: a private static
-        // method's parameters are all `object` slots, so boxing the AST-resolved value type would
-        // mismatch the `object` argument actually loaded (StackUnexpected). Mirrors EmitAsyncStubMethod.
-        var paramTypes = methodBuilder.GetParameters();
-
-        // Copy parameters to state machine fields (static methods start params at index 0).
-        for (int i = 0; i < parameters.Count; i++)
-        {
-            var field = smBuilder.GetVariableField(parameters[i].Name.Lexeme);
-            if (field != null)
-            {
-                il.Emit(OpCodes.Dup);  // Keep state machine reference on stack
-                il.Emit(OpCodes.Ldarg, i);
-
-                if (i < paramTypes.Length && paramTypes[i].ParameterType.IsValueType)
-                {
-                    il.Emit(OpCodes.Box, paramTypes[i].ParameterType);
-                }
-
-                il.Emit(OpCodes.Stfld, field);
-            }
-        }
-
-        // Return the state machine (which implements IEnumerable<object>)
-        il.Emit(OpCodes.Ret);
-    }
+    // EmitGeneratorInstanceStubMethod / EmitGeneratorStaticStubMethod were folded into the shared
+    // EmitIteratorMethodStub (ILCompiler.IteratorStubs.cs, #1126).
 }
